% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouch-package.R
\docType{package}
\name{blouch-package}
\alias{blouch-package}
\alias{blouch}
\title{The 'blouch' package.}
\description{
Bayesian Linear Ornstein-Uhlenbeck models for Comparative Hypotheses (BLOUCH)
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.4. https://mc-stan.org
}
