% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouchOUPredict.setup.v1.R
\name{blouchOUPredict.setup.v1}
\alias{blouchOUPredict.setup.v1}
\title{This is the R setup file for blouchOUPredict_v1.stan
It includes the function sigma.X.estimate, taken from the R Pacakge Slouch (Kopperud et al. 2020).}
\usage{
blouchOUPredict.setup.v1(trdata, trdata.fos, names.traits, classical)
}
\arguments{
\item{trdata}{Data formatted by make.treedata function from the R Pacakge treeplyr - only extant species}

\item{trdata.fos}{Data formatted by make.treedata function from the R Pacakge treeplyr - extant and fossil}

\item{names.traits}{Vector of trait names}

\item{classical}{Numeric denoting whether classical regression (=1) or inverse regression (=0) should be used in prediction}
}
\value{
An object of class list for use in blouchOUPredict_v1.stan
}
\description{
This is the R setup file for blouchOUPredict_v1.stan
It includes the function sigma.X.estimate, taken from the R Pacakge Slouch (Kopperud et al. 2020).
}
