% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouchOUReg.setup.v1.R
\name{blouchOUReg.setup.v1}
\alias{blouchOUReg.setup.v1}
\title{This is the R setup file for blouchOUReg_v1.stan
It includes the function sigma.X.estimate, taken from the R Pacakge Slouch (Kopperud et al. 2020).}
\usage{
blouchOUReg.setup.v1(trdata, names.traits)
}
\arguments{
\item{trdata}{Data formatted by make.treedata function from the R Pacakge treeplyr}

\item{names.traits}{Vector of trait names}
}
\value{
An object of class list for use in blouchOUReg_v1.stan
}
\description{
This is the R setup file for blouchOUReg_v1.stan
It includes the function sigma.X.estimate, taken from the R Pacakge Slouch (Kopperud et al. 2020).
}
