% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouchOU.setup.v1.R
\name{blouchOU.setup.v1}
\alias{blouchOU.setup.v1}
\title{This is the R setup file for blouchOU_v1.stan}
\usage{
blouchOU.setup.v1(trdata, names.traits)
}
\arguments{
\item{trdata}{Data formatted by make.treedata function from the R Pacakge treeplyr}

\item{names.traits}{Vector of trait names}
}
\value{
An object of class list for use in blouchOU_v1.stan
}
\description{
This is the R setup file for blouchOU_v1.stan
}
